Drop obsolete environment "prod"
When compiled as phar, env phar is in use.
When running an uncomiled tenside, we are definately a developer working
on the code base and therefore "dev" shall get used.
Env "test" is retained for the unit tests.

Therefore, "prod" was obsolete all the time and finally gets removed for
good.