Export and use the PHP cli path from the cli binary itself
While searching the PHP cli path we use the first path which
produces a valid result. However some hosters are re-writing
exec paths, so the wrong (and possibly non-existent) path to the
PHP cli ends in our configuration.

Instead of taking the path as-is we should rather use the
path the PHP cli provides (PHP_BINARY).